Actualizar lógica de etiquetas para incluir gráfica de radar
